be/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
black==25.12.0
certifi==2026.1.4
cfgv==3.5.0
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distlib==0.4.0
Django==6.0
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
filelock==3.20.2
identify==2.6.15
idna==3.11
iniconfig==2.3.0
injector==0.23.0
isort==7.0.0
librt==0.7.7
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
pytokens==0.3.0
PyYAML==6.0.3
requests==2.32.5
ruff==0.14.10
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
virtualenv==20.35.4
